inventory-microservice/src/main/java/com/metao/book/product/application/service/ProductCategoriesService.java
package com.metao.book.product.application.service;

import com.metao.book.product.domain.ProductCategoriesInterface;
import com.metao.book.product.domain.ProductCategoryEntity;
import com.metao.book.product.domain.ProductEntity;
import com.metao.book.product.domain.ProductId;
import com.metao.book.product.domain.ProductRepository;
import java.util.Set;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@Transactional
@RequiredArgsConstructor
public class ProductCategoriesService implements ProductCategoriesInterface {

    private final ProductRepository productRepository;

    @Override
    public Set<ProductCategoryEntity> getProductCategories(ProductId productId) {
        return productRepository.findById(productId).map(ProductEntity::getCategories).orElseThrow();
    }
}
